requirements.txt
python-dotenv==1.0.1
requests==2.32.3
isodate==0.6.1
yt-dlp
tenacity==9.0.0
flask==3.0.0
